Implement zeros()

+++ integration_tests/test_numpy_01.py
@@ -1,10 +1,10 @@
 # This test handles various aspects of local arrays using the `numpy.empty()`
 # function
-from ltypes import f32
+from ltypes import f64
 from numpy import empty
 
 def test_local_arrays():
-    a: f32[16]
+    a: f64[16]
     a = empty(16)
     i: i32
     for i in range(16):
